Changed float to double to decrease number of casts

+++ src/algorithms/pso_seeker.h
@@ -21,9 +21,9 @@
 
 class Particle {
 public:
-  float azimuth, elevation;
-  float velocity_azimuth, velocity_elevation;
-  float best_azimuth, best_elevation;
+  double azimuth, elevation;
+  double velocity_azimuth, velocity_elevation;
+  double best_azimuth, best_elevation;
   float best_magnitude;
 
   Antenna &antenna;
@@ -35,7 +35,7 @@ public:
 
   void random();
 
-  float compute(float azimuth, float elevation);
+  float compute(double azimuth, double elevation);
 
   void update();
 };
@@ -45,7 +45,7 @@ public:
 class PSO {
 public:
   std::vector<Particle> particles;
-  float global_best_azimuth, global_best_elevation;
+  double global_best_azimuth, global_best_elevation;
   float global_best_magnitude;
   int n_particles;
   Antenna &antenna;
